Updated test architecture to use one set of tests for all crypto tests
